feat: Update Navbar component to use react-router-dom's useNavigate hook
This commit modifies the Navbar component to use the useNavigate hook from the react-router-dom library. It adds a handleClick function that prevents the default behavior of the anchor tag and uses the navigate function to redirect to the "/contact" route when the "Contact" link is clicked. This change improves the navigation functionality of the Navbar component.